Load push cert and key once
