+++ app/services.go
@@ -1,7 +1,11 @@
 package app
 
 import (
-	"crypto/tls"
+	"crypto/rsa"
+	"crypto/x509"
+	"encoding/asn1"
+	"encoding/pem"
+	"errors"
 	"io/ioutil"
 
 	pushcertificate "github.com/RobotsAndPencils/buford/certificate"
@@ -21,6 +25,8 @@ import (
 	"github.com/micromdm/micromdm/enroll"
 	"github.com/micromdm/micromdm/management"
 	"github.com/micromdm/micromdm/workflow"
+
+	"golang.org/x/crypto/pkcs12"
 )
 
 // setupServices uses the values from the config to set up the various components
@@ -34,6 +40,7 @@ func setupServices(config *Config, logger log.Logger) (*serviceManager, error) {
 	sm.setupWorkflowDatastore()
 	sm.setupCertificateDatastore()
 
+	sm.loadPushCerts()
 	sm.setupPushService()
 
 	sm.setupCommandService()
@@ -56,6 +63,7 @@ type serviceManager struct {
 	ApplicationDatastore application.Datastore
 
 	PushService *push.Service
+	PushServiceCert
 
 	CommandService    command.Service
 	ManagementService management.Service
@@ -69,16 +77,82 @@ type serviceManager struct {
 	err    error
 }
 
+type PushServiceCert struct {
+	*x509.Certificate
+	PrivateKey interface{}
+}
+
+func (s *serviceManager) loadPushCerts() {
+	if s.err != nil {
+		return
+	}
+	if s.APNS.PrivateKeyPath == "" {
+		pkcs12Data, err := ioutil.ReadFile(s.APNS.CertificatePath)
+		if err != nil {
+			s.err = err
+			return
+		}
+		s.PushServiceCert.PrivateKey, s.PushServiceCert.Certificate, err =
+			pkcs12.Decode(pkcs12Data, s.APNS.PrivateKeyPass)
+		if err != nil {
+			s.err = err
+			return
+		}
+	} else {
+		pemData, err := ioutil.ReadFile(s.APNS.CertificatePath)
+		if err != nil {
+			s.err = err
+			return
+		}
+
+		pemBlock, _ := pem.Decode(pemData)
+
+		s.PushServiceCert.Certificate, err = x509.ParseCertificate(pemBlock.Bytes)
+		if err != nil {
+			s.err = err
+			return
+		}
+
+		pemData, err = ioutil.ReadFile(s.APNS.PrivateKeyPath)
+		if err != nil {
+			s.err = err
+			return
+		}
+
+		pemBlock, _ = pem.Decode(pemData)
+		s.PushServiceCert.PrivateKey, err = x509.ParsePKCS1PrivateKey(pemBlock.Bytes)
+		if err != nil {
+			s.err = err
+			return
+		}
+	}
+}
+
+var oidASN1UserID = asn1.ObjectIdentifier{0, 9, 2342, 19200300, 100, 1, 1}
+
+func topicFromCert(cert *x509.Certificate) (string, error) {
+	for _, v := range cert.Subject.Names {
+		if v.Type.Equal(oidASN1UserID) {
+			return v.Value.(string), nil
+		}
+	}
+
+	return "", errors.New("Could not find Push Topic (UserID OID) in certificate")
+}
+
 func (s *serviceManager) setupEnrollmentService() {
 	if s.err != nil {
+		return
+	}
+	pushTopic, err := topicFromCert(s.PushServiceCert.Certificate)
+	if err != nil {
+		s.err = err
 		return
 	}
 	// TODO: clean up order of inputs. Maybe pass *SCEPConfig as an arg?
 	// but if you do, the packages are coupled, better not.
 	s.EnrollmentService, s.err = enroll.NewService(
-		s.APNS.CertificatePath,
-		s.APNS.PrivateKeyPass,
-		s.APNS.PrivateKeyPath,
+		pushTopic,
 		s.Enrollment.CACertPath,
 		s.SCEP.RemoteURL,
 		s.SCEP.Challenge,
@@ -157,31 +231,10 @@ func (s *serviceManager) setupPushService() {
 	if s.err != nil {
 		return
 	}
-	var tlscert tls.Certificate
-	var err error
-	if s.APNS.PrivateKeyPath == "" {
-		// note that buford does validity checks where
-		// our direct certificate parsing does not
-		cert, key, err := pushcertificate.Load(
-			s.APNS.CertificatePath,
-			s.APNS.PrivateKeyPass,
-		)
-		if err != nil {
-			s.err = err
-			return
-		}
-		tlscert = pushcertificate.TLS(cert, key)
-	} else {
-		tlscert, err = tls.LoadX509KeyPair(
-			s.APNS.CertificatePath,
-			s.APNS.PrivateKeyPath,
-		)
-		if err != nil {
-			s.err = err
-			return
-		}
-	}
-	client, err := push.NewClient(tlscert)
+	client, err := push.NewClient(pushcertificate.TLS(
+		s.PushServiceCert.Certificate,
+		s.PushServiceCert.PrivateKey.(*rsa.PrivateKey),
+	))
 	if err != nil {
 		s.err = err
 		return
